Add a feed validation tool (PP-2184) (#107)
Add a tool to validate a feed against the Palace Pydantic OPDS models.
Downside of this is that it requires us to pull the whole palace.manager
package in, which brings with it a lot of dependencies. Might see if I
can sort this out.

Note: I don't love that this brings in all the dependencies from
circulation, 99% of them are unneeded here, but it is really nice to be
able to use the OPDS models defined there to do validation. Not sure the
best thing to do here.

You can see all the deps being brought in via the massive update to
`poetry.lock`. Maybe it doesn't really matter, as long as everything has
a reasonably easy to install wheel.