Fix poetry package cache in GitHub workflows (#143)
## Summary

This PR fixes the poetry package cache configuration in the GitHub
Actions workflows.

This should fix the issue seen in
https://github.com/ThePalaceProject/palace-tools/pull/142, where linting
was failing in CI.

## Details

Because the various poetry installation invocations were using the same
cache, the first one to hit would cache its installation, and the other
would restore the cache with the wrong packages. It looks like this
happened in https://github.com/ThePalaceProject/palace-tools/pull/141.