Add test infrastructure to the palace-tools project. (#150)
As the palace-tools project grows and increasingly becomes part of staff
workflows, it would be good to begin adding tests as we add new features
to ensure code quality. This PR adds pytest to the toml file, the github
ci workflow and adds a smoke test to confirm that the test
infrastructure is working.